Mapstruct lib added. Implemented DTOs and Mappers, CartService logic reimplemented

+++ MMSApplication/src/main/java/com/javaschool/onlineshop/dao/ProductDAO.java
@@ -8,7 +8,7 @@ public interface ProductDAO {
 
     List<Product> findAll();
 
-    Product getProductById(int id);
+    Product getProductById(Long id);
 
     void addProduct(Product product);
 
@@ -23,6 +23,4 @@ public interface ProductDAO {
     List<Product> findAllActiveProductsByBrandOrModel(String brand, String category);
 
     List<Product> findAllActiveProductsByName(String productName);
-
-
 }
